Hack out openmp dependence.
The reason for trying this was the number of functions showing up in v tune
for which it wasn't clear where they are coming from and wheter they actively
contribute to our running time.

At the end this didn't change much ... so the mystery where these function
come from and how hard they are hitting us remains open.