Proper phi wrap-around.
This relies on extra Config::g_MaxHitsConsidered hits copiped at the end of each BunchOfHits (this was implemented already before). This way vgather index vector can be kept unchanged through all FindHitsMinimizeCopy() and prefetching does not have to correct for the phi shift-back.

Still need to implement mask for vgather for loading hits to account for different number of hits being considered for each track. Now that we do propagation at the end of the layer loop we could try to group tracks that require significantly more hits (or a much larger phi range) into the same processing slot / matriplex.